Create Using Global Variables.cpp
Using global variables indiscriminately is generally considered poorprogramming practice.Global variables can be assigned values in any function and can contain an unpredictable state, especially if different function modules are programmed by different programmers in a team. An elegant way of having the result of multiplication in main() as required in Listing 3.3 is to have MultiplyNumbers()return thatresult to main.