This Basic Code enables the use to see the sizes of various types in bytes and is specific to there platform: compiler, OS, and hardware.
This Section of the book as thought me the difference between each 
types.

Personal Notes.
-Unsigned int types can hold more data than that of signed due to the 
extra block needed to signify if the 'int' is positive or negative
- this example has specifically been built to hold 'using namesapce 
std;' as a local var rather than global even though i know that it is 
much easier to declare this as a global var. For good programing 
practise.