Notes.
There are cases where the type of a variable is apparent given the 
initialization value it is being assigned. For example, if a variable is 
being initialized with the value true, the type of the variable can be 
best estimated as bool. In C++11,you have the option to not explicitly 
specify the type when you use keyword autoinstead.