The program displays the binary result of the various operations.
Interesting is to note the output in the event the two supplied integers 
are the same. 
The operators ==,>=, and <= produce identical results.

The fact that the output of equality and relational operators is binary 
makes these perfectly suited to using them in statements that help in 
decision-making and as loop condition expressions that ensure a loop 
executes only so long as the condition evaluates to true (1).