The program demonstrates how sizeof is capable of returning the size of an array of 100 integers in bytes, which is 400 bytes.
The program also demonstrates that the size of each element is 4 bytes. 
sizeof can be quite useful when you need to dynamically allocate memory 
for N objects, especially of a type created by yourself. You would use 
the result of the sizeof operation in determining the amount of memory 
occupied by each object and then dynamically allocate using the operator 
new.