Lesson 9.10 Commit
Note: Use the singleton pattern only where absolutely necessary, keeping 
future growth of the application and its features in perspective. Note 
that the very feature that it restricts creation of multipleinstances 
can become an architectural bottleneck when a usecase comes up that 
needs multiple instances of the class. For example, if the project was 
growing from modeling a nation to modeling the United Nations that is 
currently represented by 192 nations with 192 presidents, clearly we 
have an architectural problem in the making.